requirements.txt
Click==7.0
Flask==1.0.2
ItsDangerous==1.0.0
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.3.0
pkg-resources==0.0.0
pytesseract==0.2.5
Wand==0.4.4
Werkzeug==0.14.1
